Replacing still more uses of unsigned.

+++ FindPanel/OgreTextFindComponents/NSOutlineView/OgreOutlineViewAdapter.m
@@ -297,7 +297,7 @@
     return [_outlineView window];
 }
 
-- (unsigned)numberOfDescendantsInSelection:(BOOL)inSelection
+- (NSUInteger)numberOfDescendantsInSelection:(BOOL)inSelection
 {
     return -1;  // indeterminate
 }
